Modify error handling when user_id is invalid (#62)
いいね削除時にユーザーが不正な場合のハンドリング

+++ rr-et/app/Http/Controllers/CommentController.php
@@ -35,7 +35,11 @@ class CommentController extends Controller
     
     public function destroy(int $id)
     {
-        $comment = Comment::findOrfail($id);
+        $comment = Comment::findOrfail($id)->with('user')->with('script');
+
+        if (($comment->user_id !== Auth::user()->id) || Auth::user()->role === 1) {
+            return redirect()->route('scripts.index')->with('errors', 'ユーザーが不正です。');
+        }
 
         $comment->delete();
